+++ comms/include/ax25.h
@@ -42,11 +42,13 @@ is (8*x)/5. As a result, the maximum number of bytes in a frame is [(8*x) + (8*x
 #define AX25_DEST_ADDR_POSITION AX25_START_FLAG_BYTES
 #define AX25_SRC_ADDR_POSITION (AX25_DEST_ADDR_POSITION + AX25_DEST_ADDR_BYTES)
 #define AX25_CONTROL_BYTES_POSITION (AX25_SRC_ADDR_POSITION + AX25_SRC_ADDR_BYTES)
+/* Index of PID for a mod 8 frame (ex: U frame) which only has one control byte */
 #define AX25_MOD8_PID_POSITION (AX25_CONTROL_BYTES_POSITION + AX25_MOD8_CONTROL_BYTES)
+/* Index of PID for a mod 128 frame (ex: our I Frames or S frames) which have two control bytes */
 #define AX25_MOD128_PID_POSITION (AX25_CONTROL_BYTES_POSITION + AX25_MOD128_CONTROL_BYTES)
 #define AX25_INFO_FIELD_POSITION (AX25_MOD128_PID_POSITION + AX25_PID_BYTES)
 #define AX25_I_FRAME_FCS_POSITION (AX25_INFO_FIELD_POSITION + AX25_INFO_BYTES)
-#define AxX25_S_FRAME_FCS_POSITION AX25_INFO_FIELD_POSITION
+#define AX25_S_FRAME_FCS_POSITION AX25_INFO_FIELD_POSITION
 #define AX25_U_FRAME_FCS_POSITION (AX25_MOD8_PID_POSITION + AX25_PID_BYTES)
 
 #define AX25_FLAG 0x7E
@@ -59,9 +61,6 @@ is (8*x)/5. As a result, the maximum number of bytes in a frame is [(8*x) + (8*x
 
 #define MAX_U_FRAME_CMD_VALUE 2
 
-#ifdef __cplusplus
-extern "C" {
-#endif
 typedef struct {
   uint8_t data[AX25_MINIMUM_I_FRAME_LEN];
   uint16_t length;
@@ -86,6 +85,10 @@ typedef enum { U_FRAME_CMD_CONN, U_FRAME_CMD_DISC, U_FRAME_CMD_ACK } u_frame_cmd
 
 extern ax25_addr_t cubesatCallsign;
 extern ax25_addr_t groundStationCallsign;
+
+#ifdef __cplusplus
+extern "C" {
+#endif
 
 /**
  * @brief adds ax.25 headers onto telemetry being downlinked and stores the length of the packet in az25Data->length
